JAMES-2521 Limit CLI mailbox dependencies
mailbox-adapter is mixing:
 - adapters between mailbox and server concepts
 - JMX apis

This mix was harmfull, as we ended up binding server and mailbox implementation in the CLI

Splitting both concepts in different modules solves the issue.
