+++ mailbox/api/src/main/java/org/apache/james/mailbox/model/SerializableQuota.java
@@ -17,75 +17,67 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.store.mail.model;
+package org.apache.james.mailbox.model;
 
 import java.io.Serializable;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Function;
 
 import org.apache.james.core.quota.QuotaValue;
 
 import com.google.common.base.MoreObjects;
 
-public class SerializableQuotaValue<T extends QuotaValue<T>> implements Serializable {
-
-    public static <U extends QuotaValue<U>> SerializableQuotaValue<U> valueOf(Optional<U> input) {
-        return new SerializableQuotaValue<>(input.orElse(null));
-    }
+public class SerializableQuota<T extends QuotaValue<T>> implements Serializable {
 
     public static final long UNLIMITED = -1;
 
-    private final Long value;
-
-    public SerializableQuotaValue(T value) {
-        this(encodeAsLong(value));
+    public static <U extends QuotaValue<U>> SerializableQuota<U> newInstance(Quota<U> quota) {
+        return newInstance(quota.getUsed(), quota.getLimit());
     }
 
-    SerializableQuotaValue(Long value) {
-        this.value = value;
+    public static <U extends QuotaValue<U>> SerializableQuota<U> newInstance(U used, U max) {
+        return new SerializableQuota<>(
+            new SerializableQuotaValue<>(used),
+            new SerializableQuotaValue<>(max)
+        );
+    }
+
+    private final SerializableQuotaValue<T> max;
+    private final SerializableQuotaValue<T> used;
+
+    private SerializableQuota(SerializableQuotaValue<T> used, SerializableQuotaValue<T> max) {
+        this.max = max;
+        this.used = used;
+    }
+
+    public Long encodeAsLong() {
+        return max.encodeAsLong();
+    }
+
+    public Long getUsed() {
+        return Optional.ofNullable(used).map(SerializableQuotaValue::encodeAsLong).orElse(null);
     }
 
     @Override
     public boolean equals(Object o) {
-        if (o instanceof SerializableQuotaValue<?>) {
-            SerializableQuotaValue<?> that = (SerializableQuotaValue<?>) o;
-            return Objects.equals(value, that.value);
+        if (o instanceof SerializableQuota<?>) {
+            SerializableQuota<?> that = (SerializableQuota<?>) o;
+            return Objects.equals(max, that.max) &&
+                Objects.equals(used, that.used);
         }
         return false;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(value);
+        return Objects.hash(max, used);
     }
 
     @Override
     public String toString() {
         return MoreObjects.toStringHelper(this)
-            .add("value", value)
+            .add("max", max)
+            .add("used", used)
             .toString();
-    }
-
-    private static <U extends QuotaValue<U>> Long encodeAsLong(U quota) {
-        if (quota.isLimited()) {
-            return quota.asLong();
-        }
-        return UNLIMITED;
-    }
-
-    public Long encodeAsLong() {
-        return value;
-    }
-
-    public Optional<T> toValue(Function<Long, T> factory, T unlimited) {
-        Long longValue = encodeAsLong();
-        if (longValue == null) {
-            return Optional.empty();
-        }
-        if (longValue == UNLIMITED) {
-            return Optional.of(unlimited);
-        }
-        return Optional.of(factory.apply(longValue));
     }
 }
