+++ pom.xml
@@ -1339,6 +1339,11 @@
             </dependency>
             <dependency>
                 <groupId>${james.groupId}</groupId>
+                <artifactId>james-server-mailbox-jmx</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${james.groupId}</groupId>
                 <artifactId>james-server-mailetcontainer-api</artifactId>
                 <version>${project.version}</version>
             </dependency>
