+++ server/container/guice/mailbox/pom.xml
@@ -49,6 +49,10 @@
             <artifactId>james-server-guice-utils</artifactId>
         </dependency>
         <dependency>
+            <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-mailbox-adapter</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.inject</groupId>
             <artifactId>guice</artifactId>
         </dependency>
